Change AriaAnnouncer to remove the announced text from the DOM
