+++ .changeset/grumpy-guests-suffer.md
@@ -1,0 +1,5 @@
+---
+"@salt-ds/core": patch
+---
+
+Fix AriaAnnouncer not removing announced text from the DOM
